Done 09.1.4

+++ Labs/Lab9/09.1.3.py
@@ -3,7 +3,7 @@ import random
 
 def bar_code(values : list) -> None:
     max_value = max(values)
-    asteriscs_for_unit = 40/max_value
+    asteriscs_for_unit = int(40/max_value)
     
     for value in values:
         print("*" * int(asteriscs_for_unit * value))
